Revert "Google Sign up temporary fix + footer/banner tweaks"
