+++ components/Policies/PolicyPage.module.css
@@ -3,57 +3,6 @@
   font-family: "Nunito";
 }
 
-.subHeading {
-  font-size: 1.5rem;
-  font-weight: bold;
-  margin-top: 0;
-  margin-left: 0.2rem;
-  margin-bottom: 0.4rem;
-  letter-spacing: -0.625px;
-  text-align: center;
-  width: 100%;
-}
-
-.subHeading2 {
-  font-size: 1rem;
-  margin-top: 0;
-  margin-left: 0.2rem;
-  margin-bottom: 0.4rem;
-  letter-spacing: -0.625px;
-  text-align: center;
-  width: 100%;
-}
-
-.subHeading2Underlined {
-  font-size: 1rem;
-  text-decoration: underline;
-  margin-top: 0;
-  margin-left: 0.2rem;
-  margin-bottom: 0.4rem;
-  letter-spacing: -0.625px;
-  text-align: center;
-  width: 100%;
-}
-
-.subHeading2Bold {
-  font-size: 1rem;
-  font-weight: bold;
-  margin-top: 0;
-  margin-left: 0.2rem;
-  margin-bottom: 2rem;
-  letter-spacing: -0.625px;
-  text-align: center;
-  width: 100%;
-}
-
-hr.bottomBorder {
-  border-bottom: 2px solid rgba(0, 0, 0, 100%);
-  width: 50%;
-  margin-top: 0;
-  margin-left: 25%;
-  margin-right: 25%;
-}
-
 /* ---------- Policy Links --------------*/
 .policyContent button {
   display: flex;
