+++ components/ProfilePage/StyledProfileComponents.tsx
@@ -22,6 +22,16 @@ export const ContactInfoRow = styled(Row)`
   font-weight: 500;
 `
 
+export const Banner = styled(Row)`
+  font-family: Nunito;
+  font-size: 25px;
+  text-align: center;
+  justify-content: center;
+  padding: 0.75rem;
+  color: white;
+  background-color: #ff8600;
+`
+
 export const UserIcon = styled(Image).attrs(props => ({
   alt: "",
   src: props.src || "/profile-individual-icon.svg",
