+++ components/TestimonyCard/TestimonyItem.tsx
@@ -128,13 +128,7 @@ export const TestimonyItem = ({
                 width={50}
               />
             </Internal>
-            {/**
-             * visually-hidden until delete is ready at a later stage
-             */}
-            <Internal
-              className={`styles.link, visually-hidden`}
-              href={billLink}
-            >
+            <Internal className={styles.link} href={billLink}>
               <Image
                 className="px-2 align-self-center"
                 src="/delete-testimony.svg"
