+++ components/auth/TermsOfService.module.css
@@ -21,8 +21,3 @@
 .link {
   text-decoration: none;
 }
-
-.subscribe {
-  text-align: center;
-  text-decoration: none;
-}
