+++ components/auth/TermsOfServiceModal.tsx
@@ -3,8 +3,6 @@ import { Button, Col, Modal, Row, Stack } from "../bootstrap"
 import styles from "./TermsOfService.module.css"
 import SVG from "react-inlinesvg"
 import { External } from "components/links"
-import { NavLink } from "../Navlink"
-
 
 export default function TermsOfServiceModal({
   show,
@@ -101,14 +99,6 @@ export default function TermsOfServiceModal({
                     Cancel
                   </Button>
                 </Col>
-              </Row>
-              <Row>
-                <NavLink
-                href="https://cdn.forms-content.sg-form.com/fc8a7d49-d903-11ed-9e53-c2519c5b83a4"
-                other={{ className: `${styles.subscribe}`, target:"_blank", rel:"noopener noreferrer" }}
-                >
-                click here to subscribe to our newsletter
-                </NavLink>
               </Row>
             </Col>
           </Stack>
