+++ components/auth/UserSignUpModal.tsx
@@ -174,19 +174,9 @@ export default function UserSignUpModal({
                   </Button>
                 )}
 
-
-
-                {
-                /* 
-                Hiding social sign on buttons for now until we can get them working #TODO
-                See #1080
-
                 <Divider className="px-4">or</Divider>
 
-                <SocialSignOnButtons onComplete={onHide} /> 
-                */
-                }
-
+                <SocialSignOnButtons onComplete={onHide} />
               </Stack>
             </Form>
           </Col>
