--- components/shared/StyledSharedComponents.tsx
@@ -1,12 +1,0 @@
-import styled from "styled-components"
-import { Container, Image, Col, Row } from "../bootstrap"
-
-export const Banner = styled(Row)`
-  font-family: Nunito;
-  font-size: 25px;
-  text-align: center;
-  justify-content: center;
-  padding: 0.75rem;
-  color: white;
-  background-color: #ff8600;
-`
