+++ functions/src/email/partials/footer.handlebars
@@ -7,7 +7,8 @@
     <!-- Unsubscribe -> add link to User's profile - notifications model -->
     <!--   or backend function that handles Unsubscribe -->
     Unsubscribe |
-    <a href="https://mapletestimony.org/policies/copyright" target="_blank" rel="noopener noreferrer">Copyright</a>
+    <a href="https://mapletestimony.org/policies/terms-of-service" target="_blank" rel="noopener noreferrer">Terms &
+      Conditions</a>
     |
     <a href="https://mapletestimony.org/policies/privacy-policy" target="_blank" rel="noopener noreferrer">Privacy and
       Community Guidelines</a>
