+++ public/locales/en/footer.json
@@ -9,7 +9,7 @@
     "team": "Our Team"
   },
   "links": {
-    "browseBills": "Browse Bills",
+    "browsePolicies": "Policies",
     "accountSignIn": "Sign In",
     "accountSignOut": "Sign Out",
     "accountProfile": "Profile",
@@ -21,10 +21,11 @@
     "resourcesLegislators": "Find your Legislators",
     "teamNEU": "Northeastern University NuLawLab",
     "teamCFB": "Code for Boston",
+    "teamBC": "Boston College Law School",
     "teamHarvard": "Harvard Berkman Klein Center"
   },
   "legal": {
     "disclaimer": "MAPLE is an independent project, not affiliated with the Massachusetts legislature",
-    "TOS": "Our Privacy Policy & Terms of Service"
+    "TOS": "Our Policies and Terms of Service"
   }
 }
