+++ public/terms-of-service.md
@@ -1,3 +1,5 @@
+# Terms of Service 
+
 + ## Copyright
 + ---
 + - If you believe that any content residing or accessible on or through MAPLE infringes a copyright, please contact us at mapletestimony@gmail.com. 
@@ -5,3 +7,8 @@
 + - This work is licensed under CC BY 4.0
 + - This work is licensed under Attribution 4.0 International
 
+---
+
++ ## Newsletter
++ ---
++ - We may send newsletters via email to MAPLE members. You can unsubscribe from these newsletters at any time by using the “Unsubscribe” link in the emails themselves. We never share with or sell your email to third party providers.
